use a static feature bundle
Always using the latest feature bundle will end up enabling any new
feature that may not be compatible with the rest of the code. That is
why the features are not enabled by default.

The signatures feature has gone stable with perl 5.36, so it has been
added to the feature bundle. This is not compatible with code using
prototypes, as they have overlapping syntax. One of the modules in this
dist is using prototypes, so enabling the latest feature bundle will
break the code on 5.36.

Replace the dynamic feature bundle selection with a static feature
bundle. Since dist defines a minimum perl version of 5.14, enabling the
5.14 feature bundle seems most appropriate.